+++ mlxtend/preprocessing/__init__.py
@@ -6,9 +6,9 @@ from .transformer import TransformerObj
 from .mean_centering import MeanCenterer
 from .shuffle import shuffle_arrays_unison
 from .scaling import minmax_scaling
-from .scaling import standardizing
+from .scaling import standardize
 from .dense_transformer import DenseTransformer
 
 
 __all__ = ["MeanCenterer", "shuffle_arrays_unison",
-           "minmax_scaling", "standardizing", "DenseTransformer"]
+           "minmax_scaling", "standardize", "DenseTransformer"]