+++ mlxtend/preprocessing/mean_centering.py
@@ -1,6 +1,7 @@
 # Sebastian Raschka 08/20/2014
 # mlxtend Machine Learning Library Extensions
 
+import numpy as np
 from .transformer import TransformerObj
 
 
@@ -8,35 +9,29 @@ class MeanCenterer(TransformerObj):
     """
     Class for column centering of vectors and matrices.
 
-    Keyword arguments:
-        X: NumPy array object where each attribute/variable is
-           stored in an individual column.
-           Also accepts 1-dimensional Python list objects.
-
-    Class methods:
-        fit: Fits column means to MeanCenterer object.
-        transform: Uses column means from `fit` for mean centering.
-        fit_transform: Fits column means and performs mean centering.
-
-    The class methods `transform` and `fit_transform` return a new numpy array
-    object where the attributes are centered at the column means.
+    Attributes
+    -----------
+    col_means : numpy.ndarray [n_columns]
+        NumPy array storing the mean values for centering after fitting
+        the MeanCenterer object.
 
     """
     def __init__(self):
         self.col_means = None
 
     def transform(self, X):
-        self._get_array(X)
-        # centering
-        for i in range(self.ary.shape[0]):
-            self.ary[i] -= self.col_means
-        return self.ary
+        X_tr = np.copy(self._get_array(X))
+        X_tr = np.apply_along_axis(func1d=lambda x: x - self.col_means, axis=1, arr=X_tr)
+        return X_tr
 
     def fit(self, X):
-        self._get_array(X)
-        self.col_means = self.ary.mean(axis=0)
+        self.col_means = self._get_array(X).mean(axis=0)
         return self
 
     def fit_transform(self, X):
         self.fit(X)
         return self.transform(X)
+
+    def _get_array(self, X):
+        X_fl = X.astype('float64')
+        return X_fl